+++ IsekaiMod/Content/Classes/IsekaiProtagonist/Archetypes/GodEmporer.cs
@@ -33,23 +33,25 @@ namespace IsekaiMod.Content.Classes.IsekaiProtagonist.Archetypes
             var GodEmperorArchetype = Helpers.CreateBlueprint<BlueprintArchetype>("GodEmperorArchetype", bp => {
                 bp.LocalizedName = Helpers.CreateString($"GodEmperorArchetype.Name", "God Emperor");
                 bp.LocalizedDescription = Helpers.CreateString($"GodEmperorArchetype.Description", "Rather than wandering aimlessly, collecting harems, or defeating demon lords, "
-                    + "some protagonists decide to become gods. They sacrifice their special powers and sneak attack to gain powerful auras and a journey toward godhood.");
-                bp.LocalizedDescriptionShort = Helpers.CreateString($"GodEmperorArchetype.Description", "Rather than wandering aimlessly, collecting harems, or defeating demon lords, "
-                    + "some protagonists decide to become gods. They sacrifice their special powers and sneak attack to gain powerful auras and a journey toward godhood.");
+                    + "some protagonists decide to become gods. They sacrifice their special powers and sneak attack to gain powerful auras and a journey towards godhood.");
+                bp.LocalizedDescriptionShort = Helpers.CreateString($"GodEmperorArchetype.DescriptionShort", "Rather than wandering aimlessly, collecting harems, or defeating demon lords, "
+                    + "some protagonists decide to become gods. They sacrifice their special powers and sneak attack to gain powerful auras and a journey towards godhood.");
                 bp.IsArcaneCaster = false;
                 bp.IsDivineCaster = true;
                 bp.RemoveFeatures = new LevelEntry[] {
-                    Helpers.LevelEntry(1, IsekaiProtagonistProficiencies),
-                    Helpers.LevelEntry(3, SpecialPowerSelection),
+                    Helpers.LevelEntry(1, SneakAttack, IsekaiProtagonistProficiencies),
+                    Helpers.LevelEntry(3, SneakAttack, SpecialPowerSelection),
                     Helpers.LevelEntry(5, SneakAttack),
-                    Helpers.LevelEntry(7, SpecialPowerSelection),
-                    Helpers.LevelEntry(9, SneakAttack, FriendlyAuraFeature),
-                    Helpers.LevelEntry(10, OverpoweredAbilitySelection2),
+                    Helpers.LevelEntry(7, SneakAttack, SpecialPowerSelection),
+                    Helpers.LevelEntry(9, SneakAttack, SpecialPowerSelection, FriendlyAuraFeature),
+                    Helpers.LevelEntry(10),
+                    Helpers.LevelEntry(11, SneakAttack, SpecialPowerSelection),
                     Helpers.LevelEntry(12, BeachEpisodeSelection),
                     Helpers.LevelEntry(13, SneakAttack, SpecialPowerSelection),
-                    Helpers.LevelEntry(15, OverpoweredAbilitySelection2, OtherworldlyStamina),
+                    Helpers.LevelEntry(15, SneakAttack, OtherworldlyStamina),
                     Helpers.LevelEntry(17, SneakAttack, SpecialPowerSelection),
-                    Helpers.LevelEntry(20, SpecialPowerSelection, TrueMainCharacter),
+                    Helpers.LevelEntry(19, SneakAttack, SpecialPowerSelection),
+                    Helpers.LevelEntry(20, TrueMainCharacter),
                 };
                 bp.AddFeatures = new LevelEntry[] {
                     Helpers.LevelEntry(1, GodEmperorProficiencies),
@@ -59,7 +61,7 @@ namespace IsekaiMod.Content.Classes.IsekaiProtagonist.Archetypes
                     Helpers.LevelEntry(10, DarkAuraFeature),
                     Helpers.LevelEntry(12, SiphoningAuraFeature),
                     Helpers.LevelEntry(15, GodlyVessel),
-                    Helpers.LevelEntry(18, OverpoweredAbilitySelection2),
+                    Helpers.LevelEntry(18),
                     Helpers.LevelEntry(20, Godhood),
                 };
             });